Add files via upload
During this project, we developed a comprehensive dashboard to address the inquiries of stakeholders. By leveraging the capabilities of Power BI and Excel, we analyzed various aspects of the two hotels' performance. Notably, the revenue of both hotels did not exhibit a significant upward trend upon investigation. Consequently, we determined that there is no immediate need to expand the parking capacity for either property. On a positive note, the average daily rate demonstrated a slight but consistent upward trend, indicating a promising aspect of their performance.